requirements.txt
# List python packages to include in your build here
aiohttp
aiosqlite
customtkinter
discord
ollama
pillow
python-dotenv
requests